+++ CHANGELOG.md
@@ -61,6 +61,7 @@ GAMSPy 1.0.0
   - Fix missing links in api reference.
   - Set default problem type as MIP instead of LP.
   - Allow UniverseAlias in assignments.
+  - Add performance ci step to check model generation time difference.
   - Update gamspy_base and gamsapi to 47.5.0.
 - Documentation
   - Add a warning about the manipulation of records via .records.